Backend_Inventory/requirements.txt
asgiref==3.11.0
Django==4.2.27
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
PyJWT==2.10.1
sqlparse==0.5.5
channels==4.0.0 
daphne
PyMySQL==1.1.2
